Addressed comments of Carlos.
Updated readme.
Added Foriegn keys.
Imporved the exception handling.
Addressed all comments.

+++ parking/src/parking.py
@@ -4,14 +4,12 @@ import json
 import django
 from parking.models import TariffPlan, carDataDetails, ParkingLevel 
 from parking.src.Ticket import Ticket
-from parking.src.common import parkingExceptionsDict, LevelWithTheSameName, ParkingLevelDoesntExist, NoSpaceLeftInParking, CarWithSameNumExist
+from parking.src.common import CustomException
 
 class ParkingStation:
     '''Class for Parking station to add or remove cars,
 	   also for extending the Levels, updating tariff plans'''
-    def __init__(self):
-        '''Constructor method'''
-        self.jsonDec = json.decoder.JSONDecoder()
+    jsonDec = json.decoder.JSONDecoder()
 	
     def __displayLevels(self):
         ''' Returns the existing parking levels in the parking station.'''
@@ -24,84 +22,71 @@ class ParkingStation:
 		
     def deleteLevel(self,level_name):
         ''' Delete or close a specific parking level from the station.'''
-        ParkingLevel.objects.filter(level_num=level_name).delete()
+        ParkingLevel.objects.get(level_num=level_name).delete()
         
     def addLevel(self,level_name,total_spots):
         '''Add or extend a new parking llevel to the station.'''
-        if len(ParkingLevel.objects.filter(level_num=level_name)) > 0:
+        if ParkingLevel.objects.filter(level_num=level_name).count() > 0:
             # Check whether a parking level with the same name already exist.
-            return parkingExceptionsDict["LevelWithTheSameName"]
+            return CustomException("LevelWithTheSameName").value
 
         # Update the free spots and occupied spots based on total spots.
         free_spots  = list(range(1,int(total_spots)+1))
         occupied_spots = []
         
         # Create a new level with the details provided.
-        level = ParkingLevel(level_num=level_name,total_spots=total_spots,free_spots=json.dumps(list(free_spots)),occupied_spots=json.dumps(list(occupied_spots)))
-        level.save()
-        return self.__displayLevels()
+        try:
+            level = ParkingLevel(level_num=level_name,total_spots=total_spots,free_spots=json.dumps(list(free_spots)),occupied_spots=json.dumps(list(occupied_spots)))
+            level.save()
+        except:
+            return CustomException("FailedCreateParkingLevel").value
+        else:
+            return self.__displayLevels()
 
     def __getAvailableLocation(self):
         '''Retrieves and returns an available spot by searching from all the available Levels.'''
         location = None
-        if len(ParkingLevel.objects.all()) == 0:
+        levels = ParkingLevel.objects.all()
+        if levels.count() == 0:
             # Check whether atleast one parking level exist.
-            raise ParkingLevelDoesntExist
-        for level in ParkingLevel.objects.all():
+            raise CustomException("ParkingLevelDoesntExist")
+        for level in levels:
             # Searching for free spots in each parking level.
-            free_spots = list(self.jsonDec.decode(level.free_spots))
+            free_spots = list(ParkingStation.jsonDec.decode(level.free_spots))
             if len(free_spots) > 0:
                 # Assign a parking spot in this specific level if free spots available.
                 assignedSpot = free_spots[-1]
                 location = level.level_num + "_" + str(assignedSpot)
                 break
         # Raise an exception if no space left in any of the parking level.
-        if location is None: raise NoSpaceLeftInParking
+        if location is None: raise CustomException("NoSpaceLeftInParking")
         return location
         
-    def __checkCarNo(self,car_num):
-        '''Check whether car with the name exist in the parking slot.'''
-        try:
-            # Retrieving the car details with the name.
-            car = carDataDetails.objects.get(carno=car_num)
-        except carDataDetails.DoesNotExist:
-            # Do nothing and try to assign a spot for the car 
-            # if car number not exist in the parking.
-            pass
-        else:
-            # Raise an exception if car already there with the same name.
-            raise CarWithSameNumExist
 
     def addCar(self, car_num, plan_name):
-        '''Add/Allocate a available spot to a car in the parking.'''
+        '''Add/Allocate an available spot to a car in the parking.'''
         try:
-            tariff = TariffPlan.objects.get(plan=plan_name)
+            if TariffPlan.objects.filter(tariff_plan=plan_name).count() == 0:
+                raise CustomException("TariffPlanDoesntExist")
             inTime=time.time()
             # Get the available free spot.
             location = self.__getAvailableLocation()
             # Check the car existance in the parking.
-            self.__checkCarNo(car_num)
-            # Create a new car object in the database.
-            newCar = carDataDetails.objects.create(carno=car_num,tariff_plan=plan_name,inTime=inTime,location=location)
-            newCar.save()
-            # Assign a spot to the car.
-            self.__assignSpot(location)
-            # Create Ticket with fare details.
-            ticket = Ticket(location)
-            receipt = ticket.printTicket("Entry")
-        except ParkingLevelDoesntExist:
-            # Exception if no parking level exist.
-            receipt = parkingExceptionsDict["ParkingLevelDoesntExist"]
-        except NoSpaceLeftInParking:
-            # Exception if no free spots available in the parking.
-            receipt = parkingExceptionsDict["NoSpaceLeftInParking"]
-        except CarWithSameNumExist:
-            # Exception if there is car with the same name in parking.
-            receipt = parkingExceptionsDict["CarWithSameNumExist"]
-        except TariffPlan.DoesNotExist:
-            # Exception if there is no tariff plan defined.
-            receipt = parkingExceptionsDict["TariffPlanDoesntExist"]
-        finally:
+            if carDataDetails.objects.filter(carno=car_num).count() == 0:
+                # Create a new car object in the database.
+                tariff_plan = TariffPlan.objects.get(tariff_plan=plan_name)
+                newCar = carDataDetails.objects.create(carno=car_num,tariff_plan=tariff_plan,inTime=inTime,location=location)
+                newCar.save()
+                # Assign a spot to the car.
+                self.__assignSpot(location)
+                # Create Ticket with fare details.
+                ticket = Ticket(location)
+                receipt = ticket.printTicket("Entry")
+            else:
+                raise CustomException("CarWithSameNumExist")
+        except Exception as e:
+            return e.value
+        else:
             return receipt
             
     def removeCar(self, location):
@@ -109,23 +94,20 @@ class ParkingStation:
 		    from the parking space and make it available for next cars. 
 		'''
         try:
+            # Delete the car entry if exist otherwise raise an exception.
+            if carDataDetails.objects.filter(location=location).count() == 0:
+                raise CustomException("NoCarFound")
+
             # Create Ticket with fare details.
             ticket = Ticket(location)
             receipt = ticket.printTicket("Exit")
-            # Delete the car object from database.
             carDataDetails.objects.get(location=location).delete()
+            # Delete the car object from database.
             # Unassign or free the occipied spot of the car.
             self.__unAssignSpot(location)
-        except TariffPlan.DoesNotExist:
-            # Exception if tariffplan doesn't exist.
-            receipt = parkingExceptionsDict["TariffPlanDoesntExist"]
-        except carDataDetails.DoesNotExist:
-            # Exception if the spcified location if already free.
-            receipt = parkingExceptionsDict["LocationEmpty"]
-        except carDataDetails.MultipleObjectsReturned:
-            # Exception if there is multiple cars allocated with the same location.
-            receipt = parkingExceptionsDict["MultipleCarsWithSameLocation"]
-        finally:
+        except Exception as e:
+            return e.value
+        else:
             return receipt
 		
     def displayCars(self):
@@ -145,13 +127,13 @@ class ParkingStation:
         # Retrieve level object.
         level = ParkingLevel.objects.get(level_num=location.split("_")[0])
         # Get free spots available in that level.
-        free_spots = list(self.jsonDec.decode(level.free_spots))
+        free_spots = list(ParkingStation.jsonDec.decode(level.free_spots))
         assignedSpot = location.split("_")[1]
         # Remove a assigned spot from free spots and update the db.
         free_spots.remove(int(assignedSpot))
         level.free_spots = json.dumps(list(free_spots))
         # Add the assigned spot in the occupied spots and update the db.
-        occupiedSlots = list(self.jsonDec.decode(level.occupied_spots))
+        occupiedSlots = list(ParkingStation.jsonDec.decode(level.occupied_spots))
         occupiedSlots.append(assignedSpot)
         level.occupied_spots = json.dumps(list(occupiedSlots))
         level.save()
@@ -163,10 +145,10 @@ class ParkingStation:
         level = ParkingLevel.objects.get(level_num=location.split("_")[0])
         # Un assign the spot from occupied spot and add it in the free spots.
         assignedSpot = location.split("_")[1]
-        free_spots = list(self.jsonDec.decode(level.free_spots))
+        free_spots = list(ParkingStation.jsonDec.decode(level.free_spots))
         free_spots.append(int(assignedSpot))
         level.free_spots = json.dumps(list(free_spots))
-        occupiedSlots = list(self.jsonDec.decode(level.occupied_spots))
+        occupiedSlots = list(ParkingStation.jsonDec.decode(level.occupied_spots))
         occupiedSlots.remove(assignedSpot)
         level.occupied_spots = json.dumps(list(occupiedSlots))
         level.save()
